Update `README.md` to document structs
Update the `README.md` to document that the macros will only work with struct pointers. This makes sense once the user understands what the macros are doing but otherwise it is not intuitive. An alternative to updating the documentation would be to modify the macros; at this time there does not seem to be a practical way to modify the macros to allow structs to be passed in via value.